fcl: Split fclcollision.h into header and implementation.
